separate lib for RadioDJ & DB APIs
